Implement loading.tsx and Suspense for better user experience
It is best practice to fetch data on a component level. Having said that, we can use Suspense to wait for the data on that component to be returned. This way, we don't have to wait for all of the components to render to finish loading the page.
