+++ nextjs-dashboard/app/dashboard/(overview)/loading.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import DashboardSkeleton from "../../ui/skeletons";
+
+export default function Loading() {
+  return <DashboardSkeleton />;
+}
